LPD-45310 Move playwright tests for localized object entries to their own file, rename tests and describe block
